reorganize file actionsAvailable tests
